# core-lending-credit-scoring-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-credit-scoring
  description: Credit Scoring Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/credit-score
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/scoring-cases/{scoringCaseId}:
    get:
      tags:
        - ScoringCase
      summary: Get a scoring case by ID
      operationId: getById_1
      parameters:
        - name: scoringCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScoringCaseDTO'
    put:
      tags:
        - ScoringCase
      summary: Update a scoring case
      operationId: update_1
      parameters:
        - name: scoringCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScoringCaseDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScoringCaseDTO'
    delete:
      tags:
        - ScoringCase
      summary: Delete a scoring case
      operationId: delete_1
      parameters:
        - name: scoringCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/scoring-cases/{scoringCaseId}/bureau-calls/{scoringBureauCallId}:
    get:
      tags:
        - ScoringBureauCall
      summary: Get a bureau call record by ID
      operationId: getById_4
      parameters:
        - name: scoringCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: scoringBureauCallId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScoringBureauCallDTO'
    put:
      tags:
        - ScoringBureauCall
      summary: Update a bureau call record
      operationId: update_4
      parameters:
        - name: scoringCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: scoringBureauCallId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScoringBureauCallDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScoringBureauCallDTO'
    delete:
      tags:
        - ScoringBureauCall
      summary: Delete a bureau call record
      operationId: delete_4
      parameters:
        - name: scoringCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: scoringBureauCallId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/scoring-cases:
    get:
      tags:
        - ScoringCase
      summary: List or search scoring cases
      operationId: findAll_1
      parameters:
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.loanApplicationId
          in: query
          description: Exact filter for loan application id
          required: false
          schema:
            type: string
        - name: filters.customerId
          in: query
          description: Exact filter for customer id
          required: false
          schema:
            type: string
        - name: filters.caseStatus
          in: query
          description: Exact filter for case status
          required: false
          schema:
            type: string
            enum:
              - OPEN
              - COMPLETED
              - FAILED
              - CANCELLED
        - name: filters.caseType
          in: query
          description: Exact filter for case type
          required: false
          schema:
            type: string
            enum:
              - APPLICATION
              - RE_SCORING
              - PORTFOLIO_MONITOR
              - PRE_APPROVAL
        - name: filters.remarks
          in: query
          description: Exact filter for remarks
          required: false
          schema:
            type: string
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseScoringCaseDTO'
    post:
      tags:
        - ScoringCase
      summary: Create a new scoring case
      operationId: create_1
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScoringCaseDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScoringCaseDTO'
...
  /api/v1/scoring-cases/{scoringCaseId}/bureau-calls:
    get:
      tags:
        - ScoringBureauCall
      summary: List or search bureau calls for a scoring case
      operationId: findAll_4
      parameters:
        - name: scoringCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.scoringCaseId
          in: query
          description: Exact filter for scoring case id
          required: false
          schema:
            type: string
        - name: filters.bureauName
          in: query
          description: Exact filter for bureau name
          required: false
          schema:
            type: string
            enum:
              - AXESOR
              - EQUIFAX
              - EXPERIAN
              - ASNEF
              - OTHER
        - name: filters.callDate
          in: query
          description: Exact filter for call date
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[callDate].from
          in: query
          description: Filter call date from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[callDate].to
          in: query
          description: Filter call date to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.referenceCode
          in: query
          description: Exact filter for reference code
          required: false
          schema:
            type: string
        - name: filters.isSuccess
          in: query
          description: Exact filter for is success
          required: false
          schema:
            type: boolean
        - name: filters.rawResponse
          in: query
          description: Exact filter for raw response
          required: false
          schema:
            type: string
        - name: filters.note
          in: query
          description: Exact filter for note
          required: false
          schema:
            type: string
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseScoringBureauCallDTO'
    post:
      tags:
        - ScoringBureauCall
      summary: Create a new bureau call record
      operationId: create_4
      parameters:
        - name: scoringCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScoringBureauCallDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScoringBureauCallDTO'
components:
  schemas:
...
    ScoringCaseDTO:
      required:
        - caseStatus
        - caseType
        - customerId
      type: object
      properties:
        scoringCaseId:
          type: string
          format: uuid
          readOnly: true
        loanApplicationId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        caseStatus:
          type: string
          enum:
            - OPEN
            - COMPLETED
            - FAILED
            - CANCELLED
        caseType:
          type: string
          enum:
            - APPLICATION
            - RE_SCORING
            - PORTFOLIO_MONITOR
            - PRE_APPROVAL
        remarks:
          maxLength: 1000
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
...
    ScoringBureauCallDTO:
      required:
        - bureauName
        - callDate
        - isSuccess
        - referenceCode
        - scoringCaseId
      type: object
      properties:
        scoringBureauCallId:
          type: string
          format: uuid
          readOnly: true
        scoringCaseId:
          type: string
          format: uuid
        bureauName:
          type: string
          enum:
            - AXESOR
            - EQUIFAX
            - EXPERIAN
            - ASNEF
            - OTHER
        callDate:
          type: string
          format: date-time
        referenceCode:
          maxLength: 100
          minLength: 1
          type: string
        isSuccess:
          type: boolean
        rawResponse:
          maxLength: 10000
          minLength: 0
          type: string
        note:
          maxLength: 1000
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
...
    PaginationResponseScoringCaseDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/ScoringCaseDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...
    PaginationResponseScoringBureauCallDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/ScoringBureauCallDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.